fix: moved `CommandInbox` and `CommandOutbox` to `index.ts`s
